test: move some BrowserWindow specs to the main process
